+++ .eslintrc.json
@@ -30,6 +30,7 @@
     "@typescript-eslint/no-unused-vars": "off",
     "@typescript-eslint/ban-types": "off",
     "@typescript-eslint/no-empty-interface": "off",
+
     "max-len": [
       1,
       {
