+++ .npmrc
@@ -1,0 +1,1 @@
+PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
